better structure & Makefile
